[FIX] Minor fixes

+++ app/static/scripts/datasetScripts.js
@@ -6,6 +6,27 @@ const EDIT_URL = MAIN_PAGE + "/edit/";
 
 function addDataset() {
     event.preventDefault();
+
+    const fileInput = document.getElementById('csvUpload');
+    const fileError = document.getElementById('file-error');
+
+    const datasetName = document.getElementById('name');
+    const nameError = document.getElementById('name-error');
+
+    if (!fileInput.files || fileInput.files.length === 0) {
+        fileError.style.display = 'block';
+        return;
+    } else {
+        fileError.style.display = 'none';
+    }
+
+    if (datasetName.value === '') {
+        nameError.style.display = 'block';
+        return;
+    } else {
+        nameError.style.display = 'none';
+    }
+
     const form = document.getElementById('dataset-form');
     const formData = new FormData(form);
 
@@ -24,6 +45,16 @@ function addDataset() {
 
 function editDataset(datasetId) {
     event.preventDefault();
+
+    const datasetName = document.getElementById('name');
+    const nameError = document.getElementById('name-error');
+
+    if (datasetName.value === '') {
+        nameError.style.display = 'block';
+        return;
+    } else {
+        nameError.style.display = 'none';
+    }
 
     const form = document.getElementById('dataset-form');
     const formData = new FormData(form);
@@ -45,6 +76,10 @@ function editDataset(datasetId) {
 function deleteDataset(datasetId) {
     event.preventDefault();
 
+    if(!confirm('Вы уверены?')) {
+        return;
+    }
+
     fetch(DELETE_URL + datasetId, {
         method: 'DELETE',
         redirect: 'follow'
@@ -57,4 +92,29 @@ function deleteDataset(datasetId) {
         .catch(error => {
             console.error('Error:', error);
         });
+}
+
+if (document.getElementById('csvUpload') !== null) {
+    document.getElementById('csvUpload').addEventListener('change', function () {
+        const label = document.getElementById('button-label');
+
+        if (this.files.length > 0) {
+            if (document.getElementById('file-error') !== null) {
+                document.getElementById('file-error').style.display = 'none';
+            }
+            label.textContent = this.files[0].name;
+        } else {
+            if (document.getElementById('file-error') !== null) {
+                label.textContent = 'Загрузить CSV файл';
+            } else {
+                label.textContent = 'Заменить CSV файл';
+            }
+        }
+    });
+}
+
+if (document.getElementById('name') !== null) {
+    document.getElementById('name').addEventListener('input', function () {
+        document.getElementById('name-error').style.display = 'none';
+    });
 }
